Split typechecker environment data
